+++ elena/elena_bot.py
@@ -24,6 +24,12 @@ class Elena:
         self._state['status'] = ''
         self._state['iteration_benefit'] = 0.0
         self._state['iteration_margin'] = 0.0
+        self._state['iteration_current_bid'] = 0
+        self._state['iteration_current_ask'] = 0
+        self._state['iteration_current_value'] = 0
+        self._state['iteration_current_value_USD'] = 0
+        self._state['iteration_current_margin'] = 0
+        self._state['iteration_current_total_benefit'] = 0
         self._state['left_on_asset'] = 0.0
         self._state['sleep_until'] = 0
         self._state['sell_status'] = None
@@ -31,16 +37,19 @@ class Elena:
     def iterate(self):
         if self._state['sleep_until'] < get_time():
             if not self._state['buy_order_id'] and self._state['active']:
+                self._parameters_migration()
                 buy, sell = self._estimate_buy_sel()
                 if buy > 0:
                     llog("create a new buy order")
-                    new_buy_order = self._exchange.create_buy_order(self._state['max_order'], self._state['symbol'], buy)
+                    force_buy_price = self._state['algo'] > 7
+                    new_buy_order = self._exchange.create_buy_order(self._state['max_order'], self._state['symbol'], buy, force_buy_price)
                     self._reset_state()
                     self._state['buy_order_id'] = new_buy_order['orderId']
                     self._state['buy_order'] = new_buy_order
                     self._state['buy'] = buy
                     self._state['sell'] = sell
                     self._state['status'] = 'buying'
+                    self._update_orders_status_values_and_profits(force_update_buy_order=True)
                     self._save_state()
                 else:
                     self._state['sleep_until'] = self._sleep_until(get_time(), 5)
@@ -50,7 +59,7 @@ class Elena:
                 # return --- maybe the order is executed immediately
 
             if self._state['buy_order_id'] and not self._state['sell_order_id']:
-                self._update_orders_status_values_and_profits()
+                self._update_orders_status_values_and_profits(force_update_buy_order=True)
                 buy_order = self._state['buy_order']
                 status = buy_order['status']
                 order_time = int(buy_order['time'])
@@ -76,7 +85,7 @@ class Elena:
                     self._add_buy_auto_cancel_count()
                     self._save_state()
                 else:
-                    self._state['sleep_until'] = self._sleep_until(get_time(), 2)
+                    # self._state['sleep_until'] = self._sleep_until(get_time(), 2)
                     self._save_state()
                     llog("waiting purchase")
                 # return --- maybe the order is executed immediately
@@ -115,15 +124,18 @@ class Elena:
                     order_buy_price = float(buy_order['price'])  # get the real price when bought
                     order_sell_price = float(sell_order['price'])  # get the sell price as is in the exchange
 
-                    bid, ask = self._exchange.get_order_book_first_bids_asks(self._state['symbol'])
+                    bid = self._state['iteration_current_bid']
                     minimum_profit = 1 + ((self._exchange.minimum_profit-1)/2)
                     minimum_price = order_buy_price * minimum_profit
 
-                    if bid >= minimum_price:
+                    if bid >= minimum_price:  # earning
                         if self._state['sell_auto_cancel_im_feeling_lucky_data_samples'] > 0:
                             buy, sell = self._estimate_buy_sel(data_samples=self._state['sell_auto_cancel_im_feeling_lucky_data_samples'])
                             if sell > order_sell_price:
-                                text = f'Cancel and sell at higher price, order_buy_price:{order_buy_price}, ' \
+                                buy_order_id = self._state['buy_order_id']
+                                text = f'Cancel and sell at higher price, ' \
+                                       f'order_id:{buy_order_id}, ' \
+                                       f'order_buy_price:{order_buy_price}, ' \
                                        f'order_sell_price:{order_sell_price}, new_sell_price:{sell}'
                                 llog(text)
                                 self._state['sell_status'] = text
@@ -138,12 +150,12 @@ class Elena:
                                 pass  # Cancel order and sell at bid (we get some benefit)
                         llog("Cancel order and sell at bid (we get some benefit)")
                         self._state['sell_status'] = "Cancel order and sell at bid (we get some benefit)"
-                        self._cancel_sell_order_and_create_a_new_one(bid)
-                    else:
-                        percentage_loss = (1 - bid / order_buy_price) * 100
+                        self._cancel_sell_order_and_create_a_new_one(bid, force_sell_price=True)
+                    else:  # loosing
+                        percentage_loss = self._state['iteration_current_margin'] * -1
                         absolute_loss = (order_buy_price - bid) * float(sell_order['origQty'])
                         text = f'loss%:{percentage_loss}, order_buy_price:{order_buy_price}, bid:{bid}, loss$:{absolute_loss}'
-                        llog(text)
+                        # llog(text)
                         self._state['sell_status'] = text
                         self._save_state()
 
@@ -179,6 +191,15 @@ class Elena:
         if not state.get(key) is None:
             del state[key]
 
+    def _parameters_migration(self):
+        # should be done only before a new cycle starts
+        if self._state['algo'] == 4 or self._state['algo'] == 6 or self._state['algo'] == 8:
+            self._state['algo'] = 11
+        if self._state['algo'] == 5 or self._state['algo'] == 7 or self._state['algo'] == 9:
+            self._state['algo'] = 12
+        if self._state['algo'] == 10:
+            self._state['algo'] = 13
+
     def _read_state(self):  # TODO: if it uses self._robot_filename why is not using self._state?
         fp = open(self._robot_filename, 'r')
         state = json.load(fp)
@@ -189,23 +210,31 @@ class Elena:
         self._verify_key_set_default(state, 'sell_order_id', 0)
         self._verify_key_set_default(state, 'sleep_until', 0)
         self._verify_key_set_default(state, 'sleep_until_factor', 1.5)
-        self._verify_key_set_default(state, 'buy_auto_cancel_timeout', 0)
+
+        self._verify_key_set_default(state, 'buy_auto_cancel_timeout', 5)
         self._verify_key_set_default(state, 'buy_auto_cancel_count', 0)
-        self._verify_key_set_default(state, 'sell_auto_cancel_timeout', 0)
+        self._verify_key_set_default(state, 'iteration_current_bid', 0)
+        self._verify_key_set_default(state, 'iteration_current_ask', 0)
+        self._verify_key_set_default(state, 'iteration_current_value', 0)
+        self._verify_key_set_default(state, 'iteration_current_value_USD', 0)
+        self._verify_key_set_default(state, 'iteration_current_margin', 0)
+        self._verify_key_set_default(state, 'iteration_current_total_benefit', 0)
+
+        self._verify_key_set_default(state, 'sell_auto_cancel_timeout', 5)
         self._verify_key_set_default(state, 'sell_auto_cancel_count', 0)
-        self._verify_key_set_default(state, 'sell_auto_cancel_im_feeling_lucky_data_samples', 0)
+        self._verify_key_set_default(state, 'sell_auto_cancel_im_feeling_lucky_data_samples', 10)
         self._verify_key_set_default(state, 'sell_auto_cancel_lucky_count', 0)
         self._verify_key_set_default(state, 'stop_loss_percentage_absolute', 0)
         self._verify_key_set_default(state, 'stop_loss_percentage_relative_to_accumulated_benefits', 0)
         self._verify_key_set_default(state, 'reinvest', 0)
         self._verify_key_set_default(state, 'accumulated_benefit', 0)
+        self._verify_key_set_default(state, 'accumulated_benefit_USD', 0)
         self._verify_key_set_default(state, 'accumulated_margin', 0)
         self._verify_key_set_default(state, 'sales', 0)
         self._verify_key_set_default(state, 'cycles', 0)
 
-        # bug correction
-        if state['sales'] > state['cycles']:
-            state['sales'] = state['cycles']
+        # migrations
+
         return state
 
     def _save_state(self, state=None, filename=None):
@@ -236,7 +265,7 @@ class Elena:
     def _add_sell_auto_cancel_lucky_count(self):
         self._state['sell_auto_cancel_lucky_count'] = self._state['sell_auto_cancel_lucky_count'] + 1
 
-    def _update_orders_status_values_and_profits(self):
+    def _update_orders_status_values_and_profits(self, force_update_buy_order=False):
         buy_order = ''
         sell_order = ''
         iteration_benefit = 0
@@ -244,16 +273,35 @@ class Elena:
         left_on_asset = 0
 
         if self._state['buy_order_id']:
-            buy_order = self._exchange.get_order(self._state['buy_order_id'], self._state['symbol'])
+            check_order = True
+            if self._state['buy_order']:
+                buy_order = self._state['buy_order']
+                status = buy_order['status']
+                check_order = not (status == OrderStatus.FILLED.value) or force_update_buy_order
+            if check_order:
+                buy_order = self._exchange.get_order(self._state['buy_order_id'], self._state['symbol'])
+                self._state['buy_order'] = buy_order
+
             if self._state['sell_order_id']:
                 sell_order = self._exchange.get_order(self._state['sell_order_id'], self._state['symbol'])
+                self._state['sell_order'] = sell_order
                 if sell_order['status'] == OrderStatus.FILLED.value:
                     iteration_benefit = float(sell_order['cummulativeQuoteQty']) - float(buy_order['cummulativeQuoteQty'])
                     iteration_margin = (iteration_benefit / float(buy_order['cummulativeQuoteQty'])) * 100
                     left_on_asset = float(buy_order['executedQty']) - float(sell_order['executedQty'])
 
-        self._state['buy_order'] = buy_order
-        self._state['sell_order'] = sell_order
+                if sell_order['status'] == OrderStatus.NEW.value:
+                    bid, ask = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+                    self._state['iteration_current_bid'] = bid
+                    self._state['iteration_current_ask'] = ask
+
+                    buy_order = self._state['buy_order']
+                    order_buy_price = float(buy_order['price'])  # get the real price when bought
+                    self._state['iteration_current_margin'] = (bid / order_buy_price) * 100 - 100
+
+                    self._state['iteration_current_value'] = bid * float(buy_order['origQty'])
+                    self._state['iteration_current_value_USD'] = self._exchange.convert_to_usd(self._state['symbol'], self._state['iteration_current_value'])
+                    self._state['iteration_current_total_benefit'] = self._state['accumulated_benefit'] + self._state['iteration_current_value'] - self._state['max_order']
 
         self._state['iteration_benefit'] = iteration_benefit
         self._state['iteration_margin'] = iteration_margin
@@ -261,6 +309,7 @@ class Elena:
 
         if iteration_benefit != 0:
             self._state['accumulated_benefit'] = self._state['accumulated_benefit'] + iteration_benefit
+            self._state['accumulated_benefit_USD'] = self._exchange.convert_to_usd(self._state['symbol'], self._state['accumulated_benefit_USD'])
             self._state['accumulated_margin'] = self._state['accumulated_margin'] + iteration_margin
             if iteration_benefit < 0:
                 llog("iteration margin <0!", self._state)
@@ -277,6 +326,8 @@ class Elena:
         json.dump(history_state, fp)
         fp.close()
 
+
+
     def _calculate_reinvest_or_losses(self):
         # re-invest
         if self._state['reinvest'] > 0 and self._state['iteration_benefit'] > 0:
@@ -285,12 +336,6 @@ class Elena:
         # losses
         if self._state['iteration_benefit'] < 0:
             self._state['max_order'] = self._state['max_order'] + self._state['iteration_benefit']
-
-        # temporary algo migration for testing
-        if self._state['algo'] == 4 or self._state['algo'] == 6:
-            self._state['algo'] = 8
-        if self._state['algo'] == 5 or self._state['algo'] == 7:
-            self._state['algo'] = 9
 
     def _create_sell_order(self, sell, force_sell_price=False):
         buy_order = self._state['buy_order']
@@ -312,7 +357,7 @@ class Elena:
             self._add_sell_auto_cancel_lucky_count()
         self._state['sell_order_id'] = 0
         self._save_state()  # TODO: review if it's necessary... if the order was canceled but the new one can't be executed in the next iteration this would be understood as a human cancelation.
-        self._create_sell_order(sell,force_sell_price)
+        self._create_sell_order(sell, force_sell_price)
 
     def _estimate_buy_sel(self, data_samples=None):
         if data_samples is None:
@@ -351,7 +396,7 @@ class Elena:
         return buy, sell
 
     def _buy_on_bid_sell_based_on_linear_regression(self, candles_df_buy_sell, sell_field, margin=0):
-        next_low, ask = self._exchange.get_order_book_first_bids_asks(self._state['symbol'])
+        next_low, ask = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
 
         next_close = self._sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=margin)
 
@@ -359,7 +404,22 @@ class Elena:
         return buy, sell
 
     def _buy_on_bid_sell_based_on_fixed_margin(self, margin):
-        buy, ask = self._exchange.get_order_book_first_bids_asks(self._state['symbol'])
+        buy, ask = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+
+        sell = buy * (1 + (margin / 100))
+
+        return buy, sell
+
+    def _buy_on_ask_sell_based_on_linear_regression(self, candles_df_buy_sell, sell_field, margin=0):
+        bid, next_low = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
+
+        next_close = self._sell_based_on_linear_regression(candles_df_buy_sell, sell_field, margin=margin)
+
+        buy, sell = Elena._ensure_sell_is_higher_than_buy_by(next_close, next_low, self._exchange.minimum_profit)
+        return buy, sell
+
+    def _buy_on_ask_sell_based_on_fixed_margin(self, margin):
+        bid, buy = self._exchange.get_cached_order_book_first_bids_asks(self._state['symbol'])
 
         sell = buy * (1 + (margin / 100))
 
@@ -400,6 +460,12 @@ class Elena:
             if algo == 9:
                 buy, sell = self._buy_on_bid_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=margin)
             if algo == 10:
+                buy, sell = self._buy_on_ask_sell_based_on_fixed_margin(margin)
+            if algo == 11:
+                buy, sell = self._buy_on_ask_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=margin)
+            if algo == 12:
+                buy, sell = self._buy_on_ask_sell_based_on_linear_regression(candles_df_buy_sell, "High", margin=margin)
+            if algo == 13:
                 buy, sell = self._buy_on_bid_sell_based_on_fixed_margin(margin)
         if algo == 3:
             buy, sell = self._buy_sell_based_on_linear_regression(candles_df_buy_sell, "Close", margin=0)